# apps/inactu-control/openapi.yaml
openapi: 3.0.3
info:
  title: Inactu Control API
  version: 0.1.0
  description: Temporary control-plane API scaffold for Inactu.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      operationId: getHealthz
      responses:
        "200":
          description: Service health
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/hash/sha256:
    post:
      summary: Compute prefixed sha256 digest for a payload string
      operationId: postSha256Hash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HashRequest"
      responses:
        "200":
          description: Digest generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashResponse"
  /v1/verify/manifest:
    post:
      summary: Parse and validate a manifest, optionally enforce a policy capability ceiling
      operationId: postVerifyManifest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyManifestRequest"
      responses:
        "200":
          description: Manifest accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyManifestResponse"
        "400":
          description: Invalid manifest or policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/verify/receipt:
    post:
      summary: Parse and verify an execution receipt hash
      operationId: postVerifyReceipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyReceiptRequest"
      responses:
        "200":
          description: Receipt accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyReceiptResponse"
        "400":
          description: Invalid receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthResponse:
      type: object
      required: [status, service, version]
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: inactu-control
        version:
          type: string
          example: 0.1.0
    HashRequest:
      type: object
      required: [payload]
      properties:
        payload:
          type: string
          example: hello
    HashResponse:
      type: object
      required: [digest]
      properties:
        digest:
          type: string
          example: sha256:2cf24dba5fb0a30e26e83b2ac5b9e29e1b161e5c1fa7425e73043362938b9824
    VerifyManifestRequest:
      type: object
      required: [manifest]
      properties:
        manifest:
          type: object
          additionalProperties: true
        policy:
          type: object
          nullable: true
          additionalProperties: true
    VerifyManifestResponse:
      type: object
      required: [name, version, artifact, capability_ceiling_ok]
      properties:
        name:
          type: string
          example: hello-skill
        version:
          type: string
          example: 0.1.0
        artifact:
          type: string
          example: sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
        capability_ceiling_ok:
          type: boolean
          example: true
    VerifyReceiptRequest:
      type: object
      required: [receipt]
      properties:
        receipt:
          type: object
          additionalProperties: true
    VerifyReceiptResponse:
      type: object
      required: [artifact, receipt_hash, valid]
      properties:
        artifact:
          type: string
        receipt_hash:
          type: string
        valid:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string